requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
faiss-cpu>=1.9.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
openai>=1.0.0
